+++ appanimals/lib/screens/animal_detail_screen.dart
@@ -93,68 +93,66 @@ class _AnimalDetailScreenState extends State<AnimalDetailScreen> {
         title: Text(widget.animal.title),
         backgroundColor: Colors.green,
       ),
-      body: SingleChildScrollView(
-        child: Padding(
-          padding: const EdgeInsetsDirectional.all(16.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Center(
-                child: Image.asset(
-                  widget.animal.imagePath,
-                  fit: BoxFit.contain,
+      body: Padding(
+        padding: const EdgeInsetsDirectional.all(16.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            Center(
+              child: Image.asset(
+                widget.animal.imagePath,
+                fit: BoxFit.contain, // ajusto la img
+              ),
+            ),
+            const SizedBox(height: 20),
+            const Text(
+              'Ingrese el ID del animal (1 a 50):',
+              style: TextStyle(fontSize: 18),
+            ),
+            const SizedBox(height: 10),
+            Container(
+              decoration: BoxDecoration(
+                border: Border.all(color: Colors.green, width: 2),
+                borderRadius: BorderRadius.circular(8),
+              ),
+              child: TextField(
+                controller: _idController,
+                keyboardType: TextInputType.number,
+                decoration: const InputDecoration(
+                  labelText: 'ID',
+                  border: InputBorder.none,
+                  contentPadding: EdgeInsets.all(10),
                 ),
+                inputFormatters: [
+                  FilteringTextInputFormatter.digitsOnly,
+                  TextInputFormatter.withFunction((oldValue, newValue) {
+                    if (newValue.text.isEmpty) return newValue;
+                    final int? value = int.tryParse(newValue.text);
+                    if (value == null || value < 1 || value > 50) {
+                      return oldValue;
+                    }
+                    return newValue;
+                  }),
+                ],
               ),
-              const SizedBox(height: 20),
-              const Text(
-                'Ingrese el ID del animal (1 a 50):',
-                style: TextStyle(fontSize: 18),
-              ),
-              const SizedBox(height: 10),
-              Container(
-                decoration: BoxDecoration(
-                  border: Border.all(color: Colors.green, width: 2),
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                child: TextField(
-                  controller: _idController,
-                  keyboardType: TextInputType.number,
-                  decoration: const InputDecoration(
-                    labelText: 'ID',
-                    border: InputBorder.none,
-                    contentPadding: EdgeInsets.all(10),
-                  ),
-                  inputFormatters: [
-                    FilteringTextInputFormatter.digitsOnly,
-                    TextInputFormatter.withFunction((oldValue, newValue) {
-                      if (newValue.text.isEmpty) return newValue;
-                      final int? value = int.tryParse(newValue.text);
-                      if (value == null || value < 1 || value > 50) {
-                        return oldValue;
-                      }
-                      return newValue;
-                    }),
-                  ],
-                ),
-              ),
-              const SizedBox(height: 20),
-              ElevatedButton(
-                onPressed: () {
-                  final id = _idController.text;
-                  if (id.isEmpty) {
-                    _showSnackBar('Por favor ingrese un ID.');
-                  } else {
-                    _fetchAnimalData(id);
-                  }
-                },
-                child: const Text('Buscar'),
-              ),
-              const SizedBox(height: 20),
-              if (_isLoading) const CircularProgressIndicator(),
-              if (_animalData != null) _buildAnimalCard(),
-            ],
-          ),
+            ),
+            const SizedBox(height: 20),
+            ElevatedButton(
+              onPressed: () {
+                final id = _idController.text;
+                if (id.isEmpty) {
+                  _showSnackBar('Por favor ingrese un ID.');
+                } else {
+                  _fetchAnimalData(id);
+                }
+              },
+              child: const Text('Buscar'),
+            ),
+            const SizedBox(height: 20),
+            if (_isLoading) const CircularProgressIndicator(),
+            if (_animalData != null) _buildAnimalCard(),
+          ],
         ),
       ),
     );
