+++ appanimals/lib/screens/buscar_screen.dart
@@ -2,11 +2,11 @@ import 'package:appanimals/widgets/botonera_navigation.dart';
 import 'package:flutter/material.dart';
 import 'package:appanimals/models/categories_model.dart';
 import 'package:appanimals/widgets/category_card.dart';
-import 'package:appanimals/screens/busqeda_id_screens/animal_detail_screen.dart';
+import 'package:appanimals/screens/animal_detail_screen.dart';
 
 class BuscarScreen extends StatelessWidget {
   BuscarScreen({super.key});
-  
+
   final List<Category> categories = [
     Category(
         title: 'cats',
