+++ appanimals/lib/widgets/botonera_navigation.dart
@@ -1,4 +1,4 @@
-import 'package:appanimals/screens/busqeda_id_screens/buscar_screen.dart';
+import 'package:appanimals/screens/buscar_screen.dart';
 import 'package:appanimals/screens/home/home_screen.dart';
 import 'package:appanimals/screens/profile_screen.dart';
 import 'package:appanimals/screens/peces/peces_list_screen.dart';
