+++ appanimals/lib/widgets/custom_drawer.dart
@@ -1,6 +1,6 @@
 import 'package:appanimals/screens/news/news_screen.dart';
 import 'package:flutter/material.dart';
-import 'package:appanimals/screens/acerca_de/acerca_de_.dart';
+import 'package:appanimals/screens/acerca_de_.dart';
 import 'package:appanimals/screens/animals_screen.dart';
 import 'package:appanimals/screens/explorar_screen.dart';
 // import 'package:provider/provider.dart';
